TCP socket events for EBPF
This implements basic socket events for TCP, it tracks
connect(2)/accept(2)/close(2) on respective sockets as well as builds the
initial snapshot of existing sockets by parsing /proc.

Sockets are expressed in quark_socket{} and correlate with processes. Since a
socket might be in more than one process, two fields are provided:
 - pid_origin, the pid that established the connection.
 - pid_last_use, the last pid that affected it, if another process did the final
   close, pid_origin remains the same, but pid_last_use points to the last one.

We stash the lookup of quark_process_lookup(pid_origin) in the event process
pointer, which *can* be NULL, and the socket itself is stashed in the socket
member.

quark_event{} might need some adjustments in the future.

The EBPF probes are a bit subpar and they currently have two issues:
 - A connection that never sent a byte doesn't produce an exit event.
   https://github.com/elastic/ebpf/issues/217
 - Exit events happen before TCP close is issued
   https://github.com/elastic/ebpf/issues/216

I don't think this is good enough and will consider it experimental until both
issues are solved, first one has a fix on the way, second one likely involves
rewriting the whole probe.

Quark sockets don't produce snapshot events, because snapshot will be removed
and the user will have to use the iterator if desired.

In order to test this, pass -S to quark-mon(8).

Co-authored-by: Nicholas Berlin <56366649+nicholasberlin@users.noreply.github.com>