Make sure we always get tgid from exit
Nasty bug introduced by me trying to be overzealous.

When an exit happens, it's undefined if it will come from the group leader or
not, it's not like exec.

We were dropping exits that didn't come from the leader, which is very evil. The
probes are careful enough to only send one exit event once the group leader
died, but that doesn't mean it comes _FROM_ the group leader, so we have to look
at tgid.

Testing this is a bit tricky, since it's pretty much random, we brute force it,
which makes the test a bit slow, but should be improved later.

We can't use pthread because we build statically, so we have to go directly
through clone(2). I lost a lot of time thinking I had to to use clone3(2)
directly, only to realize I could use clone(2) where libc already tidies up the
stack.