Zap initial snapshots. Issue #113
This removes the initial snapshot of processes that it was sent on queue bootup.
The user is now expected to use quark_queue_iter(3) to retrieve them.

This simplified the code quite a bit, and as we add more kind of events,
deciding what to send initially becomes silly. We provide decent iterators and
users rejoice.

Documentation follows on the next commit.