Implement DNS events and general packet infrastructure
This implements DNS events when bpf is available and the user requested it with QQ_DNS.

The infrastructure built around quark_packet{} aims to be generic enough so we can
use for any kind of packet.

A quark_packet{} mimics a bit of pcap with extra fields:
 - `direction` is ingress/egress.
 - `origin` tells the user the source of the packet, say it came from a DNS probe.
 - `orig_len` is the original length, before truncation, of the packet from L3
   and up.
 - `cap_len` is the truncated length, and actual length of `data`, which is
   curbed by `QUARK_MAX_PACKET`.

`data` is a bit special, it is copied out from the ebpf ringbuffer into the
raw_event. When the raw_event is processed we steal the data pointer and move it
into the local storage (qev), so the user gets a pointer to the packet without
any extra copies.

event_storage_clear() cleans up the local storage from the previous call,
freeing the last packet. We also call it in queue_close() as we now have to free
data.

DNS uses cgroup probes, and their attachment scheme is different. They cannot
be autoattached, but they can be autoloaded, in order to attach it we need to
pass a file descriptor of the root cgroup, which for now is hard coded.

Cgroup programs don't use the bpf_link storage reserved in bpf_skeleton (since
it never auto attaches), so we attach and set the link ourselves (libbpf
supports this), meaning when the bpf program is destroyed it also gets
dettached/destroyed. Thank you libbpf for not documenting anything and making me
read you for hours.