Bypass mode for ebpf, first steps
This implements a bypass mode that only works on ebpf.
In this mode, a quark_event is always QUARK_EV_BYPASS and the only member filled
is `bypass`.

The purpose of this is to make it feasible to use quark as an ebpf loader only,
letting the user tap directly into the ebpf ring.

There's no pre-processing, no caching, no allocations and so on. Bypass is
really just a direct path to access bpf's ringbuffer.