Test latest Fedora kernels from fc32-41
This improves the CI by running quark-test on all fedora versions.

krun-fedora.sh downloads the latest kernel rpm of the specified version, unpacks
and runs krun.sh on it. When Fedora updates its rpms, we end up getting the
updated version.

This also enables nested virtualization on the CI so we can run things on qemu
super fast, <1min as opposed to 5-10min on the worst case, the most important
bit is that it fixes some flaky tests that would time out.

Prompted by a question from @andrewkroh

Many thanks to @mjwolf for pointing me to the allowed GCP options in buildkite.