Add a build target for alpine, so we can ensure musl works
The main purpose of this is to add Alpine to the CI build to make sure we can
always build against a proper libc like musl.

glibc is too laxed and has a record of breakage and/or not respecting standards.

We can't really run the tests yet as the alpine kernel wasn't shipping BTF, I've
fixed it upstream, so once that makes it to the docker registry we can run the tests:

https://gitlab.alpinelinux.org/alpine/aports/-/commit/8964f46e129def63b8d62809ecad2d5bf29f5688