Rework paths, long strings and args.
Change paths from a pre-allocated 1024bytes to an allocated member.
I've considered bumping it to MAXPATHLEN but 4K is too much, each process would
end up with 12K for basically nothing. Exporing qstr would be a mistake, so just
allocate them.

We allocate it once when we pop it out of the ring and copy, that chunk then
travels all the way up, from a raw_event{} to a quark_process{}, meaning it's
now one allocation+copy, instead of one allocation+copy+copy.

Also remove args_make() (which has always been a mistake) in favor of
quark_cmdline_iter().

The change also makes qstr() pointless as the allocated buffer is passed up, so
remove that as well.