Generate README.md from quark.7. Issue #165
This uses html2markdown from
https://github.com/JohannesKaufmann/html-to-markdown to convert quark.7 into README.md.

This is what I wanted in the beginning, but the markdown converter from mandoc
is not great.

Needs a lot of massage to fix links to the HTML, and any .Em is still a broken
link because of how mandoc generates html, but it looks good enough and the
actual links do work.