Implement entity_id. Issue #155
This implements the 12 byte entity id as defined in
https://www.elastic.co/docs/reference/ecs/ecs-process#field-process-entity-id

We can't depend on dynamic linking of md or openssl, so include a standlone MIT
licensed sha256 implementation from https://github.com/ilvn/SHA256, they claim
to be formally verified, so that's something.

We now also have to link against resolv so we can get the base64 functions,
that's ok, beats already links against it.

This is a WIP as I want to make sure we compute the very same entity_id as
gosysinfo and friends.