Ensure tracefs is always mounted by changing our access pattern.
It seems the kernel doesn't automount tracefs if we do an openat + open. This
can be observed in debian9 4.9.0-13-amd64. At boot kprobe_events will not be
found since it's not mounted, if you do a simple file access to it, the kernel
magically mounts it underneath, but it doesn't do if we use openat + open, for
reasons which I frankly don't care.

So change it to issue a full open, while here improve the error message in case
we just can't open any of them, also cut the variadic fmt arg since it's not used
anyway and it was just bloating the code.