Implement file events and aggregation
This implements file events outside of QQ_BYPASS, enabled in quark-mon via -F.

We do the same dance in quark_file{} that we do with quark_socket{}, everything
is done in one block allocation and we point the different paths inside storage,
which involves some length calculation. The advantage is we don't have to manage
a ton of strings and memory remains linear.

File aggregation is a bit special, when I wrote the aggregation in the past it
was focused only at file events, and it could be decided solely by the generic
bits of raw_event{}, but file events depend on the actual inode and operation,
we have to dig deeper into the event to decide.

We aggregate only CREATE->MODIFY...->REMOVE for now, we can consider MOVE in the
future.

While the probes load on RHEL8, file creation event is not being emitted, the
probes were designed to work on later kernels, it didn't even try to load on the
patched 4.18.x of RHEL8, this is somethign to be fixed soon, MODIFY and REMOVE
events are still there.

While here, implement a trivial benchmark in quark-mon that I've kept it around
from some time.