Make quark-test valgrind friendly
This took me more time to figure out than I care to admit.

The tests that did fork+exec would blow up under valgrind with weird outout,
kprobe would even segfault.

The fork+exec was done with the binary retrieved via binpath() which reads
/proc/self/exe.

Valgrind is not a monitor that forks+execs and ptraces and whatnot, it's an
actual VM, so /proc/self/exe was pointing back to a valgrind tool! Ironically,
we would run quark-test under valgrind which in turn would try to run valgrind
in itself.

Fix it by changing how we compute our binary path, use argv[0] + realpath(3).
If you're wondering why we do the dance of forking and execing self, it's
because when we run from an initramfs, we have no other binaries to run, so
quark-test is "fully" self contained.

Part of an effort to add valgrind runs to the CI.

Co-authored-by: Nicholas Berlin <56366649+nicholasberlin@users.noreply.github.com>