Fix recursive anon structures in btf
Consider the following structure:

```
struct iov_iter {
	u8 iter_type;
	bool copy_mc;
	bool nofault;
	bool data_source;
	bool user_backed;
	union {
		size_t iov_offset;
		int last_offset;
	};
	union {
		struct iovec __ubuf_iovec;
		struct {
			union {
				/* use iter_iov() to get the current vec */
---------->			const struct iovec *__iov;
				const struct kvec *kvec;
				const struct bio_vec *bvec;
				struct xarray *xarray;
				void __user *ubuf;
			};
			size_t count;
		};
	};
	union {
		unsigned long nr_segs;
		loff_t xarray_start;
	};
};
```

If we want to get the offsetof(iov_iter, __iov) we need to go through 3
anonymous structures/unions.

This change makes it so that when we ask for iov_iter.__iov and we encounter an
anonymous structure/union in its members, we recurse into it trying to find
___iov.

Fixes tty_probes that rely on iov_iter.__iov, as a bonus it cuts the (anon) hack
we had before, now we don't need to specify anything, which implies a btfhub.c
regen in the upcoming commits.

Spotted, debugged and verified that the fix works by @nicholasberlin, many
thanks!!!

Co-authored-by: Nicholas Berlin <56366649+nicholasberlin@users.noreply.github.com>