Signal root cgroup as /, not NUL
Before this change if a process was in the root cgroup, we would get a zero
length string. If an error ocurred it would also return a zero length string.

Change it so that a successful case of a root process actually sends "/" and not
a zero length string, this way quark can know if it should signal it knows the
cgroup path or not.