Add missing trusted PID checks in probes
Trusted apps feature requires us to skip handling an event probe
if current (user-level) PID is on the trustlist, but only for
File and Network probes. This is a performance optimization - we
skip creating an event in those cases. The trusted PID map is
populated periodically by from userspace.

Some of the recently added probes were missing the checks - add them.