Support kernel 6.15.3, fix *time and improve loader
atime, mtime and ctime have been broken since kernel 6.11, so fix those. This
introduces the third handling for *time members in the kernel.

The old code was computing offsets in userland and then using it on the probe,
that's super messy and error prone. Instead, start using structure types with
the ___ suffix, with this we can have different versions of the same structure
and just cast it to the correct version, this forces libbpf to solve the
existing symbols, independently if they changed type or size. This is all
documented in:
https://nakryiko.com/posts/bpf-core-reference-guide/#handling-incompatible-field-and-type-changes

--

Fedora 41 and 42 are now at 6.15.3-100.fc41.x86_64, this change broke our probe
in 2 places, use the same trick from above to fix them:
  - kernfs_node.parent -> kernfs_node.__parent
  - tty_driver.{type,subtype} changed from a short to an enum, that involves
    also fetching the enum value and renaming the old constant.