Work around valgrind false positives
When I changed the probes to use ___ suffixed structures (see previous commit),
libbpf started to load "kernel module btfs", since it can't find candidates for
the missing struct members, therefore it searches for them in these kernel
module btfs. This change causes libbpf to issue more commands that valgrind is
not aware off, making valgrind believe the access are non-initialized, while in
fact they are.

If we force the BTF path for libbpf, it does _not_ attempt to load the
additional BTF, therefore not making the calls valgrind is confused.

Fix this by disguising a bug fix as a feature, read the environment variable
QUARK_BTF_PATH, and if present, use that as the path, if not do what we always
did.

So on the valgrind test we now force QUARK_BTF_PATH=/sys/kernel/btf/vmlinux (the
default location), just to trick libbpf into not attempting the module BTFs.

Without this, valgrind would spit 300k uninitialized accesses.