Add new func to count op func params
Since the fsnotify probe can be noisy and the parameters have changed over time, only
enable that probe when necessary. Ideally, we would test for the existence of FMODE_CREATED,
but as far as I'm aware, there's not a way to detect defines. There's another commit
(https://github.com/torvalds/linux/commit/44907d79002466049fdbb8ef15730d185e0808b4) related
to the addition of FMODE_CREATED which is detectable by testing the parameter count to the
atomic_open function pointer within struct inode_operations.

This commit adds a function to detect the commit using btf, and conditionally adds the fsnotify
probe.