Add an API for starting upcoming quark-kube-talker
This is more preparatory work for kubernetes support.

We have to be flexible, this means being able to run on an environment without
multithreading, in an environment with multithreading, and in some gc runtime
like golang.

The object is to have "something" that feeds a file descriptor with kubernetes
events, this "something" is the upcoming quark-kube-talker.

We tell quark that kubernetes is active by setting up a quark-kube-talker (or
whatever), and feeing quark_queue_open with the pipe to read from. For now,
management of the filedescriptor is done by the caller, so it must start
quark-kube-talker, get the fd, and close on its own.