Move epollfd from backend queue into quark_queue.
This is in preparation for the upcoming kubernetes big evil diff.
Before this, kprobe_queue would create its own epollfd, and bpf_queue would just
return the ringbuffer (which was also an epollfd).
With this diff, quark_queue{} has its own epollfd, so kprobe_queue and bpf_queue
register their own file descriptors in qq->epollfd.

This will make sense once kubernetes is in as the kubernetes pipe fd will be
added to qq->epollfd.