Introduce Hanson, your annoying JSON generator
This introduces Hanson and quark_event_to_ecs(), so we can convert a quark event
to the json ecs format.

Hanson is really just helpers for generating json, it doesn't keep structure, it
can't parse json, it's really just a serializer, it should be pretty fast since
we only do one allocation, done by open_memstream() in glibc that is 8KB which
is more than enough to hold one document, so that means one allocation per
document.

This is a step into publishing the ECS documents ourselves in elasticsearch.

While here, also build the map of usernames and groups, this is not fully
re-entrant as setpwent() and endpwent() share global state, but it's the best we
can do right now. When glibc does getpwent() it lazily dlopens nss_switch and
never closes, so valgrind think there is leak, so escape the database during
valgrind tests.

Co-authored-by: Nicholas Berlin <56366649+nicholasberlin@users.noreply.github.com>