Instead of hardcoding /sys/fs/cgroup, search.
There are occasions where older systems which support ebpf
have both cgroup v1 and cgroup v2 mounted. This commit adds new
functionality to parse /proc/self/mountinto and iterate all the mount
points. Each mount is opened and checked for cgroup2, and if confirmed
the iteration completes. Iteration is preferable to parsing the
filesystem field because paths can be overmounted.

Update init.c to mount cgroup2 in a non-standard location with spaces
in the directory name for testing.