Try harder to get a cgroup2 root mounting point
This looks up the path for cgroup2, in case it's not in /sys/fs/cgroup, but more
importantly, it makes sure /sys/fs/cgroup is a cgroup2, and if not, it attempts
to mount a temporary cgroup2 for loading the DNS probes.

On some systems, cgroup2 is not mounted by default, even though the kernel
supports it.

For DNS and/or every future cgroup probe to function, we need to fetch the
cgroup mounting point pass it to bpf_probes__load(). After that the fd is not
really needed, we can close and also unmount the temporary cgroup2 mount we
might have created.

Idea from Nick.