Add quark_sysinfo and more ECS fields
This enriches ecs with the following fields.
 - process.thread.capabilities.{effective,permitted}
 - process.parent.*a
 - container.runtime
 - container.name
 - container.tag
 - container.hash
 - host.hostname
 - host.name
 - host.architecture
 - host.ip
 - host.mac
 - host.boot.id
 - host.os.kernel
 - host.os.name
 - host.os.version
 - host.os.full
 - host.os.release_type

--

For capabilities, it's basically just converting the uint64_t we have into the
very-very-very large ECS representation, we build the table of capabilities to
strings manually, meaning we have to update it into the future, this is better
than having to link against libcap just to do a string conversion.

For the container.* additions, it's basically just massaging the data we already
have, so nothing new.

For host.* we need a whole lot of things, this whole lot is quark_sysinfo, we
build it on boot and then forget, in the future, we need to be able to detect
some changes and update it accordingly, for example, if ip addresses change, for
now we don't care.

We parse /etc/os-release for a bunch of host.* things.

For host.ip and host.mac we go through getifaddrs(3), this requires some
pre-processing for it to remain in the format ECS wants. The current code is
good enough for ECS, but in the future quark should have its own abstraction of
network interfaces with addresses and so on, which would not be stored as a
string.

Many thanks to Nick for the awesome review!