Add session view data to ECS. Issue #233
This adds the following fields to ECS:
 - process.group_leader.*
 - process.session_leader.*
 - process.entry_leader.*
 - process.entry_leader.entry_meta.type

group_leader and session_leader are the quark_process of proc_pgid and proc_sid,
respectively. There's a boolean that is true if they are the same as the top
level process, which is the actual process we are reporting (process.*).

The hard part was already done last year, which is to compute entry_leader, the
computation has to be opted in via QQ_ENTRY_LEADER.

There's an proecss.entry_leader.entry_meta.type which I didn't find the ECS
docs, but it's required by kibana, so I copied what the session view processor
in beats is doing.

This hasn't been tested in Kibana, so I might have missed something, but the
output looks good if I compare against cloud defend.