Add process.command_line back
Quark stores command line as an array of args, but ECS wants both an array
process.args and a flattened command line process.command_line.

Command lines can be huge, we don't want to do a realloc dance and grow the
buffer, so we must construct the string manually, for that, add a
hanson_add_escaped() that adds text verbatim, but escaped, and manually build
all the string.