Implement more cloud.* fields and quark_kube_node
This is based on the awesome work from @stanek-michal.

It adds the following fields to ECS:
 - cloud.availability_zone
 - cloud.instance.name
 - cloud.provider
 - cloud.project.name
 - cloud.region

Still missing: cloud.account.{name,id} cloud.project.id, this require out of
band kubernetes calls.

The fields added here are an attribute of the kubernetes node, this means we
have to learn which node we are running on, collect those and then enrich.

Michal's diff was collecting all nodes in the cluster, even though we ever only
match one, this code only queries and fills in the node where quark is running,
it saves lookups, a bunch of tables and it's just way simpler.

The negative point is we now need to actually know which kubernetes node we are
running on, which isn't available through the API, so we now rely on the
environment variable QUARK_NODE_NAME being set, this is easily done in
kubernetes using the downward API, something like:

```
apiVersion: v1
kind: Pod
metadata:
  name: quark-pod
spec:
    env:
    - name: QUARK_NODE_NAME
      valueFrom:
        fieldRef:
          fieldPath: spec.nodeName
```

While here, also add a `-n` option to quark-kube-talker which takes precedence
over the environment.

Another challenge is telling quark-mon to fetch the kubernetes configuration
from the environment, instead of $HOME/.kube/config, in kubernetes pod's, this
is exported via a bunch of env variables, we can now select it with:
`$ quark-mon -K ENV`
This is passed down to quark-kube-talker, if no option is passed down to
quark-kube-talker, it tries first the env and then kube/config, we still need to
pass `-K ENV` to quark-mon, since this is also how we say "enable kubernetes".

Also, vastly improve how we detect kubernetes errors, now `quark_queue_open()`
will fail if it fails to fetch the basic information is wants from kubernetes,
while before things would just go on with a warning that kubernetes failed.

Other minor changes:
 - Fix the mess with empty object in ECS
 - Make pod.labels optional, as they're supressed in case it actually has no
   labels.