Add QQ_PTRACE
Before we could only use the ptrace probes in bypass mode, and for that we were
abusing QQ_MEMFD, start peeling the stuff out of QQ_MEMFD so we can have it all
outside bypass.

PTRACE is pretty straightforward, we just bubble up what we get from the probes.

ptrace