Add QQ_TTY outside of QQ_BYPASS
This forwards tty writes as a quark event, some notes:

We should do our own tty aggregation, possibly also aggregating within a
fork+exec.

sudo traps the output of the child process and then forwards to the actual
stdout, this means that under sudo the process that does the tty write is the
parent of the actual process doing the write. This makes it difficult to filter
our own output, meaning the first time we print something we get stuck in a self
feeding loop.

In the future we can add some logic to detect we are running under
sudo, by checking the environment or maybe going up the tree and seeing if we
are the child of a sudo process.

Because of the sudo complication, I haven't added it to quark-mon(8) and/or
quark_event_dump(), but there's a test.

We don't get events for tty_writes when running on initramfs, so detect it and
escape. It's likely that we don't have a pty and our code doesn't consider an
actual tty? Needs more investigation.