Bring go bindings back to this repository, first steps. Issue #254
This is the very minimal needed to bring the bindings back to this repository.

The idea is to develop it here, and then have a makefile target to copy the
necessary things over at https://github.com/elastic/go-quark, so there we can
have the precompiled archives, both for glibc and musl.

Since we're having a go/ now, also move quark-kube-talker there.