Rewrite raw_event_process()
raw_event_process() logic was hard to follow, and things only worked because of
careful consideration of ordering and aggregation rules. This weird logic made
sense when we were doing big string copies in the end, and we had just a handful
of events.

This changes it to a much simpler logic where we process each aggregated event
separately and then enrich quark_process, it does result in some extra copies,
but it's just a dozen of words, so who cares.