Garbage collect things in the beginning of the cycle, not the end
We garbage collect things after 4 seconds, this is to allow lookups from the
user to still work.

Previously we were garbage collecting after the cycle, meaning we could
theoretically end up doing a lookup for a quark_process
mid-quark_queue_get_event(), then stash it into qev->process, and then free the
memory in gc_collect(). This would be very unlikely since we would need an event
that references the soon-to-be-collected process to arrive 4 seconds later.

Give me back my sleep, make it impossible by collecting everything before we
have a chance of referencing it.