Preliminary rule system. Issue #177
This implements the barebones of a rule system for filtering events.

A ruleset is a collection of rules.
A rule has zero or more rule_fields. Each rule_field expresses one attribute
that must match, for example process.pid is one field, and file.path is another.
A rule matches when _all_ its fields match.
A rule with no fields always matches (can be a catch-all as the last rule).
A ruleset matches the first rule it matches, meaning this is a match-first scheme.

Currently we have only two actions RA_PASS and RA_DROP, if nothinig matches,
it's PASS.
For now, only 4 rule_fields, process.pid, process.ppid, process.filename and
file.path.

The next step is to design a proper DSL ala pf.conf(5), that then compiles an
ascii line-based ruleset into a quark_ruleset.

Co-authored-by: Nicholas Berlin <56366649+nicholasberlin@users.noreply.github.com>