Add QQ_GETPID for benchmarking
I want to be able to benchmark events, for that we need something that is easy
to trigger in large volumes, hence getpid().
The probe sends a more or less full event, so it simulates "some work".
We abuse the setsid code since it's the exact same contents.

The idea is to have quark-bench binary that has some standarized benchmarks,
including rule parsing and whatnot.